Pull SKError extensions to own file
