+++ Purchases/Error Handling/ErrorUtils.swift
@@ -31,7 +31,7 @@ class ErrorUtils: NSObject {
     }
 
     /**
-     * Maps an ``BackendErrorCode`` code to a ``ErrorCode``. code. Constructs an Error with the mapped code and adds a
+     * Maps a ``BackendErrorCode`` code to a ``ErrorCode``. code. Constructs an Error with the mapped code and adds a
      * `NSUnderlyingErrorKey` in the `NSError.userInfo` dictionary. The backend error code will be mapped using
      * ``BackendErrorCode/toPurchasesErrorCode()``.
      *
@@ -46,7 +46,7 @@ class ErrorUtils: NSObject {
     }
 
     /**
-     * Maps an ``BackendErrorCode`` code to an ``ErrorCode``. code. Constructs an Error with the mapped code and adds a
+     * Maps a ``BackendErrorCode`` code to an ``ErrorCode``. code. Constructs an Error with the mapped code and adds a
      * `RCUnderlyingErrorKey` in the `NSError.userInfo` dictionary. The backend error code will be mapped using
      * ``BackendErrorCode/toPurchasesErrorCode()``.
      *
@@ -269,44 +269,6 @@ class ErrorUtils: NSObject {
 
 }
 
-private extension SKError {
-
-    func toPurchasesErrorCode() -> ErrorCode {
-        switch self.code {
-        case .unknown,
-             .cloudServiceNetworkConnectionFailed,
-             .cloudServiceRevoked,
-             .overlayTimeout,
-             .overlayPresentedInBackgroundScene:
-            return .storeProblemError
-        case .clientInvalid,
-             .paymentNotAllowed,
-             .cloudServicePermissionDenied,
-             .privacyAcknowledgementRequired:
-            return .purchaseNotAllowedError
-        case .paymentCancelled,
-             .overlayCancelled:
-            return .purchaseCancelledError
-        case .paymentInvalid,
-             .unauthorizedRequestData,
-             .missingOfferParams,
-             .invalidOfferPrice,
-             .invalidSignature,
-             .invalidOfferIdentifier:
-            return .purchaseInvalidError
-        case .storeProductNotAvailable:
-            return .productNotAvailableForPurchaseError
-        case .ineligibleForOffer,
-             .overlayInvalidConfiguration,
-             .unsupportedPlatform:
-            return .purchaseNotAllowedError
-        @unknown default:
-            return .unknownError
-        }
-    }
-
-}
-
 extension ErrorUtils {
 
     static func backendError(withBackendCode backendCode: BackendErrorCode,
@@ -355,6 +317,7 @@ private extension ErrorUtils {
              .operationAlreadyInProgressForProductError,
              .unknownBackendError,
              .invalidSubscriberAttributesError,
+             .beginRefundRequestError,
              .logOutAnonymousUserError:
             Logger.error(code.description)
         case .purchaseCancelledError,
